Split out hiredis client and protocol functionality and updated code that referenced these.
